[#32] fix: 테이블 부분 재사용 컴포넌트로 적용
- 테이블의 tableHeader props로 사용될 issueListHeader를 별도의 파일로 작성
